+++ src/Routes/Routes.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
 
-import Header from '../components/shared/Header';
+import Navigation from '@navigations/RootStackNavigator';
 import Main from '@pages/Main';
 import Test from '@pages/Test';
 import Readme from '@pages/Readme';
@@ -12,11 +12,11 @@ const Routes: React.StatelessComponent<{}> = () => {
     <BrowserRouter>
       <div>
         <div>
-          <Header />
+          <Navigation />
           <div style={{textAlign: 'center', marginTop: '20px'}}>
             <Route exact path='/' component={Main} />
             <Route path='/test' component={Test} />
-            <Route path='/readme' component={Readme } />
+            <Route path='/readme' component={Readme} />
             <Route path='/helloworld' component={HelloWorld} />
             {/* <Route path='signup' component='Signup' /> */}
             {/* <Route path='login' component='Login' /> */}
