--- src/globalStyles/globalColors.css
@@ -1,6 +1,0 @@
-:root {
-  --cooniGradient: linear-gradient(to bottom right, rgb(12, 157, 197), rgb(201, 109, 216));
-  --cooniGradientHorizontal: linear-gradient(to right, rgb(12, 157, 197), rgb(201, 109, 216));
-  --cooniTxtColor: rgb(13, 157, 197);
-  --cooniGrey: rgb(227, 227, 227);
-}
